Move MatrixSolve to its own file and fix tests
